# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TUANANHPROXY API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Ok {
  ok: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterRequest {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
  referralCode?: string | null;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface UserSession {
  id: number;
  username: string;
  balance: number;
  isAdmin: boolean;
  referralCode: string;
  referredBy?: string | null;
}

export interface Product {
  id: number;
  categoryId: number;
  categoryName: string;
  name: string;
  description?: string | null;
  price: number;
  stock: number;
  sold: number;
  imageUrl: string;
  active: boolean;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  products: Product[];
}

export interface UpsertProductRequest {
  categoryId: number;
  name: string;
  description?: string | null;
  price: number;
  stock: number;
  imageUrl: string;
  active?: boolean;
}

export interface Order {
  id: number;
  productId: number;
  productName: string;
  price: number;
  status: string;
  code: string;
  createdAt: string;
}

export interface CreateOrderRequest {
  productId: number;
}

export interface Deposit {
  id: number;
  userId: number;
  username: string;
  amount: number;
  status: string;
  note: string;
  createdAt: string;
}

export interface CreateDepositRequest {
  amount: number;
  note: string;
}

export interface Transaction {
  id: number;
  type: string;
  amount: number;
  balanceAfter: number;
  description: string;
  createdAt: string;
}

export interface AccountSummary {
  balance: number;
  totalDeposited: number;
  totalSpent: number;
  ordersCount: number;
}

export interface ReferralEntry {
  username: string;
  joinedAt: string;
  totalDeposited: number;
  commissionEarned: number;
}

export interface ReferralSummary {
  referralCode: string;
  referralLink: string;
  referredCount: number;
  totalCommission: number;
  recentReferrals: ReferralEntry[];
}

export interface AdminUser {
  id: number;
  username: string;
  balance: number;
  isAdmin: boolean;
  referralCode: string;
  referredBy?: string | null;
  totalDeposited: number;
  ordersCount: number;
  createdAt: string;
}

export interface AdjustBalanceRequest {
  /** Positive to add, negative to subtract */
  amount: number;
  reason: string;
}

export interface ProductKey {
  id: number;
  productId: number;
  keyValue: string;
  isUsed: boolean;
  usedByOrderId?: number | null;
  createdAt: string;
  usedAt?: string | null;
}

export interface ProductKeyList {
  productId: number;
  totalKeys: number;
  availableKeys: number;
  usedKeys: number;
  keys: ProductKey[];
}

export interface AddProductKeysRequest {
  /** Newline-separated list of proxy keys */
  keys: string;
}

export interface AdminStats {
  totalUsers: number;
  totalRevenue: number;
  pendingDeposits: number;
  totalOrders: number;
  recentOrders: Order[];
}