# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TUANANHPROXY API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /admin/users/{id}/balance:
    post:
      operationId: adminAdjustBalance
      tags: [admin]
      summary: Add or subtract balance for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustBalanceRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
...
  /admin/deposits/{id}/approve:
    post:
      operationId: adminApproveDeposit
      tags: [admin]
      summary: Approve a deposit (credits balance, pays referral 20%)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /admin/deposits/{id}/reject:
    post:
      operationId: adminRejectDeposit
      tags: [admin]
      summary: Reject a deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: List all products (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: adminCreateProduct
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProductRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/products/{id}:
    patch:
      operationId: adminUpdateProduct
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProductRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
  /admin/products/{id}/keys:
    get:
      operationId: adminListProductKeys
      tags: [admin]
      summary: List proxy keys in stock for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductKeyList"
    post:
      operationId: adminAddProductKeys
      tags: [admin]
      summary: Add proxy keys to product stock (one per line)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddProductKeysRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductKeyList"
  /admin/products/{id}/keys/{keyId}:
    delete:
      operationId: adminDeleteProductKey
      tags: [admin]
      summary: Delete a single (unused) proxy key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: keyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
...
components:
  schemas:
...
    Ok:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    Product:
      type: object
      required: [id, categoryId, categoryName, name, price, stock, sold, imageUrl, active]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        stock:
          type: integer
        sold:
          type: integer
        imageUrl:
          type: string
        active:
          type: boolean
    UpsertProductRequest:
      type: object
      required: [categoryId, name, price, stock, imageUrl]
      properties:
        categoryId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        stock:
          type: integer
        imageUrl:
          type: string
        active:
          type: boolean
...
    Deposit:
      type: object
      required: [id, userId, username, amount, status, note, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        amount:
          type: number
        status:
          type: string
        note:
          type: string
        createdAt:
          type: string
...
    AdminUser:
      type: object
      required: [id, username, balance, isAdmin, referralCode, totalDeposited, ordersCount, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        balance:
          type: number
        isAdmin:
          type: boolean
        referralCode:
          type: string
        referredBy:
          type: string
          nullable: true
        totalDeposited:
          type: number
        ordersCount:
          type: integer
        createdAt:
          type: string
    AdjustBalanceRequest:
      type: object
      required: [amount, reason]
      properties:
        amount:
          type: number
          description: Positive to add, negative to subtract
        reason:
          type: string
    ProductKey:
      type: object
      required: [id, productId, keyValue, isUsed, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        keyValue:
          type: string
        isUsed:
          type: boolean
        usedByOrderId:
          type: integer
          nullable: true
        createdAt:
          type: string
        usedAt:
          type: string
          nullable: true
    ProductKeyList:
      type: object
      required: [productId, totalKeys, availableKeys, usedKeys, keys]
      properties:
        productId:
          type: integer
        totalKeys:
          type: integer
        availableKeys:
          type: integer
        usedKeys:
          type: integer
        keys:
          type: array
          items:
            $ref: "#/components/schemas/ProductKey"
    AddProductKeysRequest:
      type: object
      required: [keys]
      properties:
        keys:
          type: string
          description: Newline-separated list of proxy keys
...