# lib/api-zod/src/generated/types/addProductKeysRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TUANANHPROXY API
 * OpenAPI spec version: 0.1.0
 */

export interface AddProductKeysRequest {
  /** Newline-separated list of proxy keys */
  keys: string;
}