# lib/api-zod/src/generated/types/productKey.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TUANANHPROXY API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductKey {
  id: number;
  productId: number;
  keyValue: string;
  isUsed: boolean;
  usedByOrderId?: number | null;
  createdAt: string;
  usedAt?: string | null;
}