# lib/api-zod/src/generated/types/productKeyList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TUANANHPROXY API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductKey } from "./productKey";

export interface ProductKeyList {
  productId: number;
  totalKeys: number;
  availableKeys: number;
  usedKeys: number;
  keys: ProductKey[];
}